fix scala-js build
