+++ tests/jvm/src/test/scala/antimirov/DfaTest.scala
