+++ tests/jvm/src/test/scala/antimirov/NfaTest.scala
