+++ tests/jvm/src/test/scala/antimirov/RxTest.scala
