+++ tests/shared/src/test/scala/antimirov/BitSetTest.scala
