+++ tests/shared/src/test/scala/antimirov/CharsTest.scala
