+++ tests/shared/src/test/scala/antimirov/LazyStreamTest.scala
