+++ tests/shared/src/test/scala/antimirov/LetterMapTest.scala
