+++ tests/shared/src/test/scala/antimirov/SizeTest.scala
