+++ tests/shared/src/test/scala/antimirov/Util.scala
@@ -1,10 +1,9 @@
 package antimirov
 
 import antimirov.check.Regex
-import org.scalacheck.{Arbitrary, Gen, Prop, Properties}
+import org.scalacheck.{Arbitrary, Gen, Prop}
 
 import Arbitrary.arbitrary
-import Prop.{forAllNoShrink => forAll}
 
 object Util {
 
@@ -26,46 +25,6 @@ object Util {
       println(s"$name took $t ms ($s)")
     }
     a1
-  }
-
-  import java.util.concurrent.{Callable, ForkJoinPool, FutureTask}
-
-  val pool = new ForkJoinPool()
-
-  def timed(label: => String, warnMs: Long, abortMs: Long, failOnAbort: Boolean, body: => Prop): Prop = {
-
-    val task = new FutureTask(new Callable[Prop] { def call(): Prop = body })
-    pool.execute(task)
-
-    val start = System.nanoTime()
-    val warnT = start + warnMs * 1000000L
-    val abortT = start + abortMs * 1000000L
-
-    var result: Option[Prop] = None
-    var t0 = start
-    while (!task.isDone && t0 < abortT) {
-      Thread.sleep(0L, 100000)
-      val t1 = System.nanoTime()
-      t0 = t1
-    }
-
-    if (task.isDone) {
-      result = Some(task.get)
-    } else {
-      task.cancel(true)
-    }
-
-    val now = System.nanoTime()
-    val duration = (now - start).toDouble / 1000000.0
-
-    if (result.isEmpty) {
-      val word = if (failOnAbort) "FAIL" else "STOP"
-      println(s"$word {$label} failed to finish within ${abortMs}ms")
-    } else if (now >= warnT) {
-      println("WARN {%s} took %.1fms (longer than %dms)".format(label, duration, warnMs))
-    }
-
-    result.getOrElse(!failOnAbort)
   }
 
   def stringsFromRx(r: Rx): Option[Gen[String]] =
@@ -150,46 +109,4 @@ object Util {
   val genRxAndStrs: Gen[(Rx, Set[String])] =
     Gen.zip(genRxAndStr, arbitrary[Set[String]])
       .map { case ((rx, good), maybe) => (rx, good | maybe) }
-}
-
-trait TimingProperties { self: Properties =>
-
-  import Util.timed
-
-  def enableTiming: Boolean = true
-  def failOnAbort: Boolean = true
-
-  def scale: Long = 10
-  def warnMs: Long = 50L * scale
-  def abortMs: Long = 100L * scale
-
-  def timedProp[A](name: String, ga: Gen[A])(f: A => Prop): Unit = {
-    self.property(name) =
-      if (enableTiming) {
-        forAll(ga)(a => timed(s"$name : $a", warnMs, abortMs, failOnAbort, f(a)))
-      } else {
-        forAll(ga)(f)
-      }
-    ()
-  }
-
-  def timedProp[A, B](name: String, ga: Gen[A], gb: Gen[B])(f: (A, B) => Prop): Unit = {
-    self.property(name) =
-      if (enableTiming) {
-        forAll(ga, gb)((a, b) => timed(s"$name : $a : $b", warnMs, abortMs, failOnAbort, f(a, b)))
-      } else {
-        forAll(ga, gb)(f)
-      }
-    ()
-  }
-
-  def timedProp[A, B, C](name: String, ga: Gen[A], gb: Gen[B], gc: Gen[C])(f: (A, B, C) => Prop): Unit = {
-    self.property(name) =
-      if (enableTiming) {
-        forAll(ga, gb, gc)((a, b, c) => timed(s"$name : $a : $b : $c", warnMs, abortMs, failOnAbort, f(a, b, c)))
-      } else {
-        forAll(ga, gb, gc)(f)
-      }
-    ()
-  }
 }
