+++ tests/shared/src/test/scala/antimirov/check/RegexTests.scala
